+++ HISTORY.rst
@@ -3,6 +3,14 @@
 History
 -------
 
+0.3.0 (Unreleased)
+++++++++++++++++++
+
+* Multisite support
+* Configurable default author support
+* Refactored settings
+* Fix multilanguage issues
+
 0.2.0 (2014-09-24)
 ++++++++++++++++++